pyproject.toml
[project]
name = "a2a-multiagent-host"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "google-genai>=1.9.0",
    "google-adk>=1.7.0",
    "fastapi>=0.117.1",
    "uvicorn>=0.37.0",
    "httpx>=0.28.1",
    "litellm",
]
